pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-orchestrator"
version = "0.1.0"
description = "AI orchestrator for local LLMs in air-gapped environments"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "rich>=13.0.0",
    "prompt_toolkit>=3.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "anyio>=4.0.0",
]

[project.scripts]
open-orchestrator = "open_orchestrator.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/open_orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
